requirements.txt
langgraph
langchain
langchain-google-genai
ddgs
langchain-community
langchain-core
python-dotenv
BrotliCFFI